// solves-machine/game-event-names.enum.ts
export enum GameEventsNames {
  PlayerQuickPressNotInProgress = "PlayerQuickPressNotInProgress",
  PlayerLongPress = "PlayerLongPress",
  PlayerReleaseQuickPress = "PlayerReleaseQuickPress",
  PlayerReleaseLongPress = "PlayerReleaseLongPress",
  PlayerQuickPressInProgress = "PlayerQuickPressInProgress",
  CubeTypeChanged = "CubeTypeChanged",
  SessionReset = "SessionReset",
  Tick= "Tick",
}